Add files via upload
This article describes that,“Once these wearables collect enough data to know what your normal baseline readings are, they can get very good at sensing when something’s amiss,” says Snyder. “We think that if your heart rate and skin temperature are elevated for about 2 hours, there’s a strong chance you’re getting sick.” (https://www.newscientist.com/article/2117854-smartwatches-know-youre-getting-a-cold-days-before-you-feel-ill/#ixzz6HfL1tsci) So I implemented a heartrate feature that monitors heartrate in the background and if the heartrate exceeds the average resting heartrate (76 bpm) then a notification is triggered.  Secondly the peer conectivity feature works in the background for an infinite time (so far). If a user is infected and a user was connected the user exposed user is notified.  Next steps: implementing the ui and database!  We ware doing great! :)