Add files via upload
This article describes that,“Once these wearables collect enough data to know what your normal baseline readings are, they can get very good at sensing when something’s amiss,” says Snyder. “We think that if your heart rate and skin temperature are elevated for about 2 hours, there’s a strong chance you’re getting sick.” (https://www.newscientist.com/article/2117854-smartwatches-know-youre-getting-a-cold-days-before-you-feel-ill/#ixzz6HfL1tsci) So I implemented a heart rate feature that monitors heart rate in the background and if the heart rate exceeds the average resting heart rate (76 bpm) for two hours then a notification is triggered indicating that the user is potentially sick. Secondly the peer conectivity feature works in the background for an infinite time (if Apple Watch is connected). If a user1 is infected and a user2 was connected to user1 then user2 is notified. Next steps: getting phones without an Apple Watch to stay on in the background. We are doing great! :). Just let me know what you think of the heart rate concept, if you feel we don't want it just let me know and I can remove it.